Make refactor architecture
