+++ data_network/build.gradle.kts
@@ -66,6 +66,4 @@ dependencies {
 
     // modules
     implementation(project(":common_utils"))
-    implementation(project(":search"))
-    implementation(project(":favorites"))
 }
