+++ data_network/src/main/java/ru/practicum/android/diploma/data/networkclient/api/dto/HHIndustriesResponse.kt
@@ -2,7 +2,6 @@ package ru.practicum.android.diploma.data.networkclient.api.dto
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
-import ru.practicum.android.diploma.search.domain.model.HttpStatus
 
 @Parcelize
 data class HHIndustriesResponse(
