+++ team/build.gradle.kts
@@ -62,4 +62,7 @@ dependencies {
     // modules
     implementation(project(":common_ui"))
     implementation(project(":common_utils"))
+    implementation(project(":data_network"))
+    implementation(project(":data_sp"))
+    implementation(project(":data_db"))
 }
